настройка spring security.........

+++ src/main/java/com/samsolutions/repository/UserRepository.java
@@ -1,12 +1,13 @@
 package com.samsolutions.repository;
 
 import com.samsolutions.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     User findByUsername(String username);
 }
